A: logをtxtで保存できる機能感性

+++ app/backend/utils/log/log.py
@@ -1,17 +1,17 @@
 import os
 import json
 
-
-def log_answer(file_path, answer):
+def log(file_path: str, answer):
+    if os.path.exists(file_path) == False:
+        with open(file_path, 'w') as f:
+            json.dump([answer], f)
+        return
+    
     with open(file_path, 'r') as f:
         read_data = json.load(f)
-        save_data = [read_data, answer]
+        save_data = read_data + [answer]
 
     with open(file_path, 'w') as f:
         json.dump(save_data, f)
 
-
-def log(file_path: str, answer):
-    print(file_path, answer)
-    log_answer(file_path, answer)
     return
